+++ modules/api.js
@@ -1,0 +1,36 @@
+export function getComments() {
+    return fetch("https://wedev-api.sky.pro/api/v1/Volkov_Pavel/comments", {
+        method: "GET"
+    })
+    .then((response) => {
+        if (response.status === 500) {          
+            throw new Error("Сервер сломался, попробуй позже");
+        }
+        else {
+            return response.json();
+        }      
+    })
+}
+
+export function postComment({text, name}) {
+    return fetch("https://wedev-api.sky.pro/api/v1/Volkov_Pavel/comments", {
+        method: "POST",
+        body: JSON.stringify({
+            text: text,
+            name: name,
+            forceError: true,
+        }),
+    }).then((response) => {
+        if (response.status === 400) {   
+            console.log(1);       
+            throw new Error('Имя и комментарий должны быть не короче 3 символов');          
+        }
+        else if (response.status === 500) {    
+            console.log(2);      
+            throw new Error('Сервер сломался, попробуй позже');          
+        }
+        else {
+            return response.json();
+        }
+    })
+}
